refactor: move video offset into offset package
