+++ scinoephile/media/offset/__init__.py
@@ -1,8 +1,3 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Media file operations.
-
-Package hierarchy (modules may import from any above):
-* constants / offset / probe
-* subtitles
-"""
+"""Media offset detection."""
