+++ scinoephile/media/offset/video/_video_frame_sample.py
@@ -1,0 +1,20 @@
+#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
+#  and distributed under the terms of the BSD license. See the LICENSE file for details.
+"""Sampled video frame data."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+import numpy as np
+
+
+@dataclass(frozen=True)
+class _VideoFrameSample:
+    """Sampled video frame at a timestamp."""
+
+    time: float
+    """Timestamp of the sampled frame in seconds."""
+
+    frame: np.ndarray
+    """Normalized grayscale frame data."""
